refactor: consolidate client_state methods
