--- ibc/lightclients/groth16/upgrade.go
@@ -1,24 +1,0 @@
-package groth16
-
-import (
-	"context"
-	"errors"
-
-	storetypes "cosmossdk.io/store/types"
-	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/ibc-go/v9/modules/core/exported"
-)
-
-// VerifyUpgradeAndUpdateState returns an error because it hasn't been
-// implemented yet.
-func (cs *ClientState) VerifyUpgradeAndUpdateState(
-	ctx context.Context,
-	cdc codec.BinaryCodec,
-	clientStore storetypes.KVStore,
-	upgradedClient exported.ClientState,
-	upgradedConsState exported.ConsensusState,
-	proofUpgradeClient,
-	proofUpgradeConsState []byte,
-) error {
-	return errors.New("not implemented")
-}
